# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HubWorkz Saúde API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /medicamentos:
    get:
      operationId: listMedicamentos
      tags: [medicamentos]
      summary: List medication catalog
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Medication list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medicamento"
    post:
      operationId: createMedicamento
      tags: [medicamentos]
      summary: Create a medication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicamentoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicamento"
  /medicamentos/{id}:
    get:
      operationId: getMedicamento
      tags: [medicamentos]
      summary: Get a medication by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Medication detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicamento"
  /distribuidoras:
    get:
      operationId: listDistribuidoras
      tags: [distribuidoras]
      summary: List distributors
      responses:
        "200":
          description: Distributor list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Distribuidora"
    post:
      operationId: createDistribuidora
      tags: [distribuidoras]
      summary: Create a distributor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DistribuidoraInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Distribuidora"
  /distribuidoras/{id}:
    get:
      operationId: getDistribuidora
      tags: [distribuidoras]
      summary: Get a distributor by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Distributor detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Distribuidora"
  /faturas:
    get:
      operationId: listFaturas
      tags: [faturas]
      summary: List invoices
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: convenio_id
          schema:
            type: string
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fatura"
    post:
      operationId: createFatura
      tags: [faturas]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaturaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fatura"
  /faturas/{id}:
    get:
      operationId: getFatura
      tags: [faturas]
      summary: Get an invoice by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invoice detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fatura"
    patch:
      operationId: updateFatura
      tags: [faturas]
      summary: Update an invoice
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaturaUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fatura"
  /glosas:
    get:
      operationId: listGlosas
      tags: [glosas]
      summary: List glosas
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Glosa list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Glosa"
    post:
      operationId: createGlosa
      tags: [glosas]
      summary: Create a glosa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GlosaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Glosa"
  /glosas/{id}:
    patch:
      operationId: updateGlosa
      tags: [glosas]
      summary: Update a glosa (e.g. submit recurso)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GlosaUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Glosa"
  /monitoramentos:
    get:
      operationId: listMonitoramentos
      tags: [monitoramentos]
      summary: List D30 monitoring records
      parameters:
        - in: query
          name: mes
          schema:
            type: string
        - in: query
          name: paciente_id
          schema:
            type: string
      responses:
        "200":
          description: Monitoring list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Monitoramento"
    post:
      operationId: createMonitoramento
      tags: [monitoramentos]
      summary: Create a monitoring record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitoramentoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitoramento"
  /monitoramentos/{id}:
    patch:
      operationId: updateMonitoramento
      tags: [monitoramentos]
      summary: Update a monitoring record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitoramentoUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitoramento"
  /cotacoes:
    get:
      operationId: listCotacoes
      tags: [cotacoes]
      summary: List quotations
      parameters:
        - in: query
          name: processo_id
          schema:
            type: string
      responses:
        "200":
          description: Quotation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cotacao"
    post:
      operationId: createCotacao
      tags: [cotacoes]
      summary: Create a quotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CotacaoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cotacao"
  /ai/chat:
    post:
      operationId: sendAiMessage
      tags: [ai]
      summary: Send a message to Di IA (Anthropic)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiMessageResponse"
  /ai/history:
    get:
      operationId: getAiHistory
      tags: [ai]
      summary: Get Di IA chat history
      parameters:
        - in: query
          name: clinica_id
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiMessage"
components:
  schemas:
...
    Medicamento:
      type: object
      required: [id, nome, created_at]
      properties:
        id:
          type: string
        nome:
          type: string
        principio_ativo:
          type: ["string", "null"]
        apresentacao:
          type: ["string", "null"]
        modo_uso:
          type: ["string", "null"]
        conservacao:
          type: ["string", "null"]
        registro_anvisa:
          type: ["string", "null"]
        categoria:
          type: ["string", "null"]
        created_at:
          type: string
    MedicamentoInput:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
        principio_ativo:
          type: string
        apresentacao:
          type: string
        modo_uso:
          type: string
        conservacao:
          type: string
        registro_anvisa:
          type: string
        categoria:
          type: string
    Distribuidora:
      type: object
      required: [id, nome, created_at]
      properties:
        id:
          type: string
        nome:
          type: string
        cnpj:
          type: ["string", "null"]
        responsavel:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        telefone:
          type: ["string", "null"]
        tipo:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        created_at:
          type: string
    DistribuidoraInput:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
        cnpj:
          type: string
        responsavel:
          type: string
        email:
          type: string
        telefone:
          type: string
        tipo:
          type: string
    Fatura:
      type: object
      required: [id, tipo, status, created_at]
      properties:
        id:
          type: string
        processo_id:
          type: ["string", "null"]
        paciente_id:
          type: ["string", "null"]
        convenio_id:
          type: ["string", "null"]
        tipo:
          type: string
        numero_nf:
          type: ["string", "null"]
        valor:
          type: ["number", "null"]
        status:
          type: string
        data_emissao:
          type: ["string", "null"]
        data_vencimento:
          type: ["string", "null"]
        created_at:
          type: string
    FaturaInput:
      type: object
      required: [tipo, status]
      properties:
        processo_id:
          type: string
        paciente_id:
          type: string
        convenio_id:
          type: string
        tipo:
          type: string
        numero_nf:
          type: string
        valor:
          type: number
        status:
          type: string
        data_emissao:
          type: string
        data_vencimento:
          type: string
    FaturaUpdate:
      type: object
      properties:
        status:
          type: string
        numero_nf:
          type: string
        valor:
          type: number
    Glosa:
      type: object
      required: [id, status, created_at]
      properties:
        id:
          type: string
        fatura_id:
          type: ["string", "null"]
        processo_id:
          type: ["string", "null"]
        motivo:
          type: ["string", "null"]
        valor:
          type: ["number", "null"]
        status:
          type: string
        prazo_recurso:
          type: ["string", "null"]
        recurso_texto:
          type: ["string", "null"]
        created_at:
          type: string
    GlosaInput:
      type: object
      required: [status]
      properties:
        fatura_id:
          type: string
        processo_id:
          type: string
        motivo:
          type: string
        valor:
          type: number
        status:
          type: string
        prazo_recurso:
          type: string
    GlosaUpdate:
      type: object
      properties:
        status:
          type: string
        recurso_texto:
          type: string
    Monitoramento:
      type: object
      required: [id, created_at]
      properties:
        id:
          type: string
        paciente_id:
          type: ["string", "null"]
        processo_id:
          type: ["string", "null"]
        data_contato:
          type: ["string", "null"]
        adesao:
          type: ["string", "null"]
        eventos_adversos:
          type: ["string", "null"]
        observacoes:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        created_at:
          type: string
    MonitoramentoInput:
      type: object
      properties:
        paciente_id:
          type: string
        processo_id:
          type: string
        data_contato:
          type: string
        adesao:
          type: string
        eventos_adversos:
          type: string
        observacoes:
          type: string
        status:
          type: string
    MonitoramentoUpdate:
      type: object
      properties:
        adesao:
          type: string
        eventos_adversos:
          type: string
        observacoes:
          type: string
        status:
          type: string
    Cotacao:
      type: object
      required: [id, created_at]
      properties:
        id:
          type: string
        processo_id:
          type: ["string", "null"]
        medicamento_id:
          type: ["string", "null"]
        fornecedor_tipo:
          type: ["string", "null"]
        fornecedor_id:
          type: ["string", "null"]
        valor:
          type: ["number", "null"]
        prazo_entrega:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        recomendada:
          type: boolean
        created_at:
          type: string
    CotacaoInput:
      type: object
      properties:
        processo_id:
          type: string
        medicamento_id:
          type: string
        fornecedor_tipo:
          type: string
        fornecedor_id:
          type: string
        valor:
          type: number
        prazo_entrega:
          type: string
        status:
          type: string
        recomendada:
          type: boolean
    AiMessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        clinica_id:
          type: string
        context:
          type: ["string", "null"]
    AiMessage:
      type: object
      required: [id, role, content, created_at]
      properties:
        id:
          type: string
        clinica_id:
          type: ["string", "null"]
        role:
          type: string
        content:
          type: string
        created_at:
          type: string
    AiMessageResponse:
      type: object
      required: [message, response]
      properties:
        message:
          type: string
        response:
          type: string
        tokens_used:
          type: ["integer", "null"]