# lib/api-zod/src/generated/types/aiMessageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HubWorkz Saúde API
 * OpenAPI spec version: 0.1.0
 */

export interface AiMessageInput {
  message: string;
  clinica_id?: string;
  /** @nullable */
  context?: string | null;
}