# lib/api-zod/src/generated/types/cotacaoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HubWorkz Saúde API
 * OpenAPI spec version: 0.1.0
 */

export interface CotacaoInput {
  processo_id?: string;
  medicamento_id?: string;
  fornecedor_tipo?: string;
  fornecedor_id?: string;
  valor?: number;
  prazo_entrega?: string;
  status?: string;
  recomendada?: boolean;
}