# lib/api-zod/src/generated/types/medicamentoInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HubWorkz Saúde API
 * OpenAPI spec version: 0.1.0
 */

export interface MedicamentoInput {
  nome: string;
  principio_ativo?: string;
  apresentacao?: string;
  modo_uso?: string;
  conservacao?: string;
  registro_anvisa?: string;
  categoria?: string;
}