# lib/api-zod/src/generated/types/monitoramento.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HubWorkz Saúde API
 * OpenAPI spec version: 0.1.0
 */

export interface Monitoramento {
  id: string;
  /** @nullable */
  paciente_id?: string | null;
  /** @nullable */
  processo_id?: string | null;
  /** @nullable */
  data_contato?: string | null;
  /** @nullable */
  adesao?: string | null;
  /** @nullable */
  eventos_adversos?: string | null;
  /** @nullable */
  observacoes?: string | null;
  /** @nullable */
  status?: string | null;
  created_at: string;
}